Basic Blackhole Skybox with controller for colour changing and colour transition.
Had to make a custom skybox shader in the shader graph because the colour changing functionality didn't work with the standard skybox material.

Skybox Controller prefab added just to demo with the settings I showed in the discord, The component used it likely something we put onto the game manager.

Shader graph is optimised as much as I can (with my knowledge), singular 1/x operation but that's about it. Completely negligible performance impact from both the skybox and the controller though the controller can still be further optimised.