Ripped Enemy scripts from the gamejam project
Implemented and adjusted the gamejam ai functionality, replacing goal targeting with random patrolling. Added Player sensing too. Tower functionality left in as we plan to add summonable towers (if I remember correctly, no point ripping it out if it may eventually get put back in).

Seems to behave as intended.

Still need to figure out an alternative to the use of a tag for identifying whether something is attackable.

Commented functionality that relies on things from other branches that aren't yet in main (removing health and notifying game manager).