Ghost Effect for perfect dodge implemented.
This is just an initial thing to get it into the game.

Currently its a module of its own. if we plan on having complex effects like the one we plan to have for the perfect dodge it may be worth us considering an effects system where we manage effects. It would also allow us to pool objects we use for things like the perfect dodge.